Attempted fix for uploading project proposals
On the production server, there is a problem where project proposals
can't be uploaded. This commit attempts to fix it (we can't test the fix
locally, because it only affects the server). It turns out that when using
Paperclip's has_attached_file method, the URI used for the file is an
absolute path for the whole server when it starts with a slash. Unfortunately,
this means that the server is attempting to save and retrieve files from
outside the UNL iLabs project directory.

This commit adds the UNL iLabs directory path to the start of the
has_attached_file URI in production mode to hopefully coax the server into
saving files in the right place.