Allow instructors to change a student's team
Before this commit, there was no way for an instructor to manually change the
team a student was on. The week this commit was made, there were some problems
pulling assignments from the auto-solver, so some students were incorrectly
left with no team assignments, and the instructors had no way to correct this.

This commit adds an option to change a student's team to the student edit page.
It also changes which parameters are accepted when modifying students, but only
for instructors. This means that it's possible for instructors to change a
student's team, but students can't make the change themselves, even if they
manually write HTML requests.