Front-end for assigning instructors to projects
Added a front-end for assigning instructors to projects and visa-verca. Go to
the edit form for any instructor or project and look for a link right underneath
the "Submit" button.

The new forms are functional, but I don't know if they're very intuitive or
easy to find. I'll wait to see what the client says about them first, though.