Started meeting handout
For weekly meetings, the client wants to be able to print out a page with the
names, emails, and pictures of every student, sponsor, and instructor involved
in a team. This way, at the beginning of the semester, they can print these
handouts out to make it easier for everyone to get to know each other.

This commit introduces some of the functionality to create this handout. The
meeting_handout.html.haml file is going to need to be edited to look better in
the future, probably so that it uses a printer-friendly layout and its own CSS
styling.